+++ src/joedb/io/run_interpreted_client.h
@@ -12,6 +12,45 @@
 namespace joedb
 {
  ////////////////////////////////////////////////////////////////////////////
+ inline void run_interpreted_client(Interpreted_Client &client)
+ ////////////////////////////////////////////////////////////////////////////
+ {
+  while (std::cin)
+  {
+   std::cout << "R(read), P(pull), W(write), or Q(quit)? ";
+   std::cout.flush();
+   std::string input;
+
+   if (!(std::cin >> input))
+   {
+    std::cout << '\n';
+    break;
+   }
+
+   std::cin.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+
+   if (input == "W")
+   {
+    std::cout << "connecting...";
+    std::cout.flush();
+
+    client.transaction([](Readable &readable, Writable &writable)
+    {
+     std::cout << "OK\n";
+     std::cout.flush();
+     Interpreter(readable, writable).main_loop(std::cin, std::cout);
+    });
+   }
+   else if (input == "P")
+    client.pull();
+   else if (input == "R")
+    Readonly_Interpreter(client.get_database()).main_loop(std::cin, std::cout);
+   else if (input == "Q")
+    break;
+  }
+ }
+
+ ////////////////////////////////////////////////////////////////////////////
  inline void run_interpreted_client
  ////////////////////////////////////////////////////////////////////////////
  (
@@ -33,35 +72,7 @@ namespace joedb
    file = &memory_file;
 
   Interpreted_Client client(connection, *file);
-
-  while (std::cin)
-  {
-   std::cout << "R(read), P(pull), W(write), or Q(quit)? ";
-   std::cout.flush();
-   std::string input;
-
-   if (!(std::cin >> input))
-   {
-    std::cout << '\n';
-    break;
-   }
-
-   std::cin.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
-
-   if (input == "W")
-   {
-    client.transaction([](Readable &readable, Writable &writable)
-    {
-     Interpreter(readable, writable).main_loop(std::cin, std::cout);
-    });
-   }
-   else if (input == "P")
-    client.pull();
-   else if (input == "R")
-    Readonly_Interpreter(client.get_database()).main_loop(std::cin, std::cout);
-   else if (input == "Q")
-    break;
-  }
+  run_interpreted_client(client);
  }
 }